javaserver/src/main/java/id/my/mrgsrylm/wallet/javaserver/service/impl/UserServiceImpl.java
package id.my.mrgsrylm.wallet.javaserver.service.impl;

import id.my.mrgsrylm.wallet.javaserver.model.UserModel;
import id.my.mrgsrylm.wallet.javaserver.repository.UserRepository;
import id.my.mrgsrylm.wallet.javaserver.service.UserService;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;

@Service
@Slf4j(topic = "UserService")
@RequiredArgsConstructor
public class UserServiceImpl implements UserService {
    private final UserRepository repository;

    @Override
    public UserModel getReferenceById(long id) {
        return repository.getReferenceById(id);
    }
}
